refactor(engine): group auth capabilities and shared types under capability::auth
Consolidate the bearer-token capabilities and the authorization vocabulary they
share into a single `capability::auth` domain:

- Move `bearer_token_provider` and `bearer_token_authorizer` under
  `capability::auth`, and mirror the trait-variant re-exports at
  `{local,shared}::capability::auth::<name>`.
- Merge the two per-capability `BearerToken` types into one shared credential
  (`new` for a bare token, `with_expiry` / `from_absolute_expiry` for a known
  lifetime, `from_header_value` for an Authorization header).
- Put the shared data types one-per-file under `auth/models/` (BearerToken,
  AuthorizedIdentity, AuthzDecision, DenyReason) behind a private module,
  re-exported so each has the single public path `capability::auth::<Type>`.
- Update consumers and the export-convention docs; the `#[capability]` macro
  needed no changes (it emits absolute `crate::capability::...` paths).

No behavior change.

Copilot-Session: 10d746ae-67a2-4f55-be24-202717eaadd9
