+++ rust/otap-dataflow/crates/engine/src/capability/mod.rs
@@ -7,33 +7,34 @@
 //! Extensions register capabilities via [`ExtensionCapabilities`], and
 //! node factories consume them via [`registry::Capabilities`].
 //!
-//! Capability traits are defined per-capability in submodules (e.g.,
-//! `bearer_token_provider`), with local (!Send) and shared (Send) variants
-//! re-exported from [`local::capability`](crate::local::capability) and
+//! Capability traits are defined per-capability in submodules grouped by domain
+//! (e.g., `auth::bearer_token_provider`), with local (!Send) and shared (Send)
+//! variants re-exported from [`local::capability`](crate::local::capability) and
 //! [`shared::capability`](crate::shared::capability).
 
 // Capability code is organized so every public item has exactly **one** export
 // surface — no item is reachable via two paths:
 //
 // - A capability's **data types and registration handle** are exposed only on
-//   the **scoped** surface `capability::<name>` — e.g.
-//   `capability::bearer_token_provider::{BearerTokenProvider, BearerToken,
-//   TokenStream}`. Scoping namespaces every item by its owning
-//   capability, so two capabilities can each define an identically named type
-//   (e.g. both a `BearerToken`) with no clash.
+//   the **scoped** surface under its domain. Types shared across a domain's
+//   capabilities are re-exported at the domain root, `capability::<domain>`
+//   (e.g. `capability::auth::BearerToken`); a capability's own handle and
+//   capability-specific types live in its submodule, `capability::<domain>::<name>`
+//   (e.g. `capability::auth::bearer_token_provider::{BearerTokenProvider,
+//   TokenStream}`). Their defining modules stay private, so each item has a
+//   single public path.
 //
 // - A capability's **trait variants** (the `local` `!Send` / `shared` `Send`
 //   traits the `#[capability]` macro generates) are exposed only on the
-//   **execution-model** surface, `{local,shared}::capability::<name>` — the
-//   surface extensions implement against, alongside `{local,shared}::extension`,
+//   **execution-model** surface, `{local,shared}::capability::<domain>::<name>` —
+//   the surface extensions implement against, alongside `{local,shared}::extension`,
 //   etc.
 //
 // The two surfaces don't overlap because the `#[capability]` macro emits its
 // `local`/`shared` trait modules as `pub(crate)` (an implementation detail), so
-// they are never a public path under `capability::<name>`; the traits become
-// public only through the hand-written `{local,shared}::capability` re-exports.
-// The capability module itself is a plain `pub mod`, directly exposing the data
-// types and handle — no facade or `#[path]` indirection required.
+// they are never a public path under `capability::<domain>::<name>`; the traits
+// become public only through the hand-written `{local,shared}::capability`
+// re-exports.
 //
 // - Genuinely **shared** framework infrastructure — used across capabilities
 //   rather than owned by one — is re-exported flat at `capability::` (below): the
@@ -42,8 +43,7 @@
 //   module). These are unique, collision-free vocabulary where a single short
 //   canonical path is worth more than namespacing. `registry` stays `pub`
 //   because `registry::Capabilities` is not re-exported at the root.
-pub mod bearer_token_authorizer;
-pub mod bearer_token_provider;
+pub mod auth;
 pub(crate) mod error;
 pub(crate) mod factory;
 pub mod registry;
@@ -79,11 +79,11 @@ pub trait ExtensionCapability: private::Sealed + 'static {
     const NAME: &'static str;
 
     /// The local (!Send) trait object type for this capability
-    /// (e.g., `dyn local::capability::bearer_token_provider::BearerTokenProvider`).
+    /// (e.g., `dyn local::capability::auth::bearer_token_provider::BearerTokenProvider`).
     type Local: ?Sized + 'static;
 
     /// The shared (Send) trait object type for this capability
-    /// (e.g., `dyn shared::capability::bearer_token_provider::BearerTokenProvider`).
+    /// (e.g., `dyn shared::capability::auth::bearer_token_provider::BearerTokenProvider`).
     type Shared: ?Sized + Send + 'static;
 
     /// Human-readable name used in error messages and config validation.
