+++ rust/otap-dataflow/crates/engine/src/shared/capability.rs
@@ -4,19 +4,22 @@
 //! Shared (Send) capability trait re-exports.
 //!
 //! Each production capability's shared trait variant is re-exported here
-//! under a module named after the capability (e.g.
-//! `shared::capability::bearer_token_provider::BearerTokenProvider`), mirroring
-//! the `capability::<name>` scoping so every surface is scoped the same way.
-//! Capability traits are defined by the `#[capability]` proc macro in
-//! per-capability modules under [`capability`](crate::capability). Test-only
-//! reference capabilities live under [`crate::testing::capability`] and are
-//! intentionally not re-exported here.
+//! under a module path mirroring its `capability::<domain>::<name>` scoping
+//! (e.g. `shared::capability::auth::bearer_token_provider::BearerTokenProvider`),
+//! so every surface is scoped the same way. Capability traits are defined by
+//! the `#[capability]` proc macro in per-capability modules under
+//! [`capability`](crate::capability). Test-only reference capabilities live
+//! under [`crate::testing::capability`] and are intentionally not re-exported
+//! here.
 
-/// Shared (Send) trait variant of the bearer-token-authorizer capability.
-pub mod bearer_token_authorizer {
-    pub use crate::capability::bearer_token_authorizer::shared::BearerTokenAuthorizer;
-}
-/// Shared (Send) trait variant of the bearer-token-provider capability.
-pub mod bearer_token_provider {
-    pub use crate::capability::bearer_token_provider::shared::BearerTokenProvider;
+/// Auth capabilities (shared `Send` trait variants).
+pub mod auth {
+    /// Shared (Send) trait variant of the bearer-token-authorizer capability.
+    pub mod bearer_token_authorizer {
+        pub use crate::capability::auth::bearer_token_authorizer::shared::BearerTokenAuthorizer;
+    }
+    /// Shared (Send) trait variant of the bearer-token-provider capability.
+    pub mod bearer_token_provider {
+        pub use crate::capability::auth::bearer_token_provider::shared::BearerTokenProvider;
+    }
 }
